+++ pychron/envisage/browser/table_configurer.py
@@ -16,23 +16,25 @@
 
 # ============= enthought library imports =======================
 
-from traits.api import HasTraits, List, Any, Bool, Int, Instance
+from traits.api import HasTraits, List, Any, Bool, Int, Instance, Enum
 from traits.trait_errors import TraitError
-from traitsui.api import View, Item, UItem, CheckListEditor, VGroup, Handler, HGroup
+from traitsui.api import View, Item, UItem, CheckListEditor, VGroup, Handler, HGroup, Tabbed
 import apptools.sweet_pickle as pickle
-
 #============= standard library imports ========================
 from datetime import datetime
 import os
 #============= local library imports  ==========================
 from pychron.paths import paths
 
+SIZES=(6, 8, 9, 10, 11, 12, 14, 15, 18, 24, 36)
 
 class TableConfigurerHandler(Handler):
     def closed(self, info, is_ok):
         if is_ok:
-            info.object.dump()
-            info.object.set_columns()
+            info.object.closed()
+            # info.object.dump()
+            # info.object.set_columns()
+
 
 
 class TableConfigurer(HasTraits):
@@ -40,6 +42,13 @@ class TableConfigurer(HasTraits):
     available_columns = List
     adapter = Any
     id = 'table'
+    font = Enum(*SIZES)
+
+    def closed(self):
+        self.dump()
+        self.set_columns()
+        self.set_font()
+
 
     def load(self):
         self._load_state()
@@ -60,7 +69,12 @@ class TableConfigurer(HasTraits):
             #set all available columns
             self.available_columns = acols
 
+            self._set_font(adp.font)
             self._load_state()
+
+    def _set_font(self, f):
+        s=f.pointSize()
+        self.font=s
 
     def _load_state(self):
         p = os.path.join(paths.hidden_dir, self.id)
@@ -81,6 +95,10 @@ class TableConfigurer(HasTraits):
 
                 self.columns = ncols
 
+            font = state.get('font', None)
+            if font:
+                self.font=font
+
             self._load_hook(state)
             self.set_columns()
 
@@ -95,11 +113,16 @@ class TableConfigurer(HasTraits):
                 pass
 
     def _get_dump(self):
-        obj = dict(columns=self.columns)
+        obj = dict(columns=self.columns,
+                   font=self.font)
         return obj
 
     def _load_hook(self, state):
         pass
+
+    def set_font(self):
+        if self.adapter:
+            self.adapter.font='arial {}'.format(self.font)
 
     def set_columns(self):
         # def _columns_changed(self):
@@ -135,7 +158,6 @@ def str_to_time(lp):
 class AnalysisTableConfigurer(TableConfigurer):
     id = 'analysis.table'
     limit = Int
-
 
     def _get_dump(self):
         obj = super(AnalysisTableConfigurer, self)._get_dump()
@@ -212,6 +234,7 @@ class IsotopeTableConfigurer(TableConfigurer):
         v = View(VGroup(UItem('columns',
                               style='custom',
                               editor=CheckListEditor(name='available_columns', cols=3)),
+                        Item('font'),
                         show_border=True,
                         label='Isotopes'))
         return v
@@ -224,25 +247,44 @@ class IntermediateTableConfigurer(TableConfigurer):
         v = View(VGroup(UItem('columns',
                               style='custom',
                               editor=CheckListEditor(name='available_columns', cols=3)),
+                        Item('font'),
                         show_border=True,
                         label='Intermediate'))
         return v
+
+
 
 
 class RecallTableConfigurer(TableConfigurer):
     isotope_table_configurer = Instance(IsotopeTableConfigurer, ())
     intermediate_table_configurer = Instance(IntermediateTableConfigurer, ())
     show_intermediate = Bool
+    experiment_fontsize = Enum(*SIZES)
+    measurement_fontsize = Enum(*SIZES)
+    extraction_fontsize = Enum(*SIZES)
+
+    view_names = ('experiment', 'measurement', 'extraction')
+    # def closed(self):
+    #     super(RecallTableConfigurer, self).closed()
+    #     self.experiment_view
 
     def _get_dump(self):
         obj = super(RecallTableConfigurer, self)._get_dump()
         obj['show_intermediate'] = self.show_intermediate
+        for a in self.view_names:
+            a='{}_fontsize'.format(a)
+            obj[a]=getattr(self, a)
+
         return obj
 
     def _load_hook(self, obj):
         self.show_intermediate = obj.get('show_intermediate', True)
         self.isotope_table_configurer.load()
         self.intermediate_table_configurer.load()
+
+        for a in self.view_names:
+            a='{}_fontsize'.format(a)
+            setattr(self, a, obj.get(a, 10))
 
     def dump(self):
         super(RecallTableConfigurer, self).dump()
@@ -253,10 +295,31 @@ class RecallTableConfigurer(TableConfigurer):
         self.isotope_table_configurer.set_columns()
         self.intermediate_table_configurer.set_columns()
 
+    def set_font(self):
+        self.isotope_table_configurer.set_font()
+        self.intermediate_table_configurer.set_font()
+
+    def set_fonts(self, av):
+        for a in self.view_names:
+            s=getattr(self, '{}_fontsize'.format(a))
+            av.update_fontsize(a, s)
+
     def traits_view(self):
-        v = View(VGroup(UItem('isotope_table_configurer', style='custom'),
+        main_view=VGroup(UItem('isotope_table_configurer', style='custom'),
                         HGroup(Item('show_intermediate', label='Show Intermediate Table')),
-                        UItem('intermediate_table_configurer', style='custom', enabled_when='show_intermediate')),
+                        UItem('intermediate_table_configurer', style='custom', enabled_when='show_intermediate'),
+                        label='Main')
+
+        experiment_view = VGroup(Item('experiment_fontsize'),
+                                 label='Exp.')
+        measurement_view = VGroup(Item('measurement_fontsize'),
+                                  label='Meas.')
+        extraction_view = VGroup(Item('extraction_fontsize'),
+                                  label='Ext.')
+        v = View(Tabbed(main_view,
+                        experiment_view,
+                        measurement_view,
+                        extraction_view),
                  buttons=['OK', 'Cancel', 'Revert'],
                  kind='modal',
                  title='Configure Table',